improved and documented core/Table api
